// management/frontend/src/api/schemas/collectionRecordPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionRecordPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 128
     */
  record_no: string;
  task_id?: string | null;
  /** @maxLength 128 */
  project_name?: string;
  /** @maxLength 128 */
  task_name?: string;
  /** @maxLength 128 */
  subtask_name?: string;
  /** @maxLength 128 */
  kit_name?: string;
  /** @maxLength 256 */
  capture_location?: string;
  /** @maxLength 128 */
  device_serial?: string;
  /** @maxLength 128 */
  operator_username?: string;
  /** @maxLength 256 */
  file_name?: string;
  /** @minimum 0 */
  file_size_bytes?: number;
  /** @minimum 0 */
  duration_seconds?: number;
  /** @maxLength 32 */
  status?: string;
  /** @maxLength 32 */
  qa_status?: string;
  /** @maxLength 32 */
  upload_status?: string;
  /** @maxLength 32 */
  data_status?: string;
  captured_at?: string | null;
  /** @maxLength 1000 */
  remark?: string;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}